+++ src/daisy.h
@@ -51,6 +51,7 @@
 #include "dev/dps310.h"
 #include "dev/lcd_hd44780.h"
 #include "dev/mcp23x17.h"
+#include "dev/tca9534.h"
 #include "dev/max11300.h"
 #include "dev/tlv493d.h"
 #include "dev/dotstar.h"